Correções menores

+++ src/main/java/com/kuuhaku/controller/postgresql/CardDAO.java
@@ -31,6 +31,7 @@ import com.kuuhaku.model.persistent.Card;
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
 import javax.persistence.Query;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.List;
@@ -575,7 +576,7 @@ public class CardDAO {
 		}
 	}
 
-	public static int getMeta(Class cat) {
+	public static double getMeta(Class cat) {
 		EntityManager em = Manager.getEntityManager();
 
 		Query q = em.createNativeQuery("""
@@ -592,7 +593,7 @@ public class CardDAO {
 		q.setParameter("cat", cat.name());
 
 		try {
-			return ((Double) q.getSingleResult()).intValue();
+			return ((BigDecimal) q.getSingleResult()).doubleValue();
 		} finally {
 			em.close();
 		}
